Define a single function to run in eventlisteners
I created a function for each of the actions within the eventListeners
for better design and readability. I defined a new function called
getRoundResults() which includes the logic that was in playRound().
I updated playRound() so that is uses the functions that contain the
functionality that I had within the eventListeners so I can just call
the playRound() function in the eventListeners.